projects/monylog-backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "monylog-shared-kernel",
    "monylog-shared-kernel-infra-fastapi",
    "monylog-shared-kernel-infra-database-sqla",
    "pendulum>=3.1.0",
    "nanoid>=2.0.0",
    "langchain-openai>=0.3.18",
    "sqlalchemy-fields>=0.5.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.11",
    "aiosqlite>=0.21.0",
    "faker>=35.2.0",
    "pre-commit>=4.1.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "pytest-testmon>=2.1.3",
    "alembic>=1.16.1",
]


[tool.uv.sources]
monylog-shared-kernel = { workspace = true }
monylog-shared-kernel-infra-fastapi = { workspace = true }
monylog-shared-kernel-infra-database-sqla = { workspace = true }


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/monylog"]
[tool.ruff]
extend = "../../pyproject.toml"
